+++ src/Controller/Admin/AdminDashboardController.php
@@ -6,6 +6,7 @@ use Kuusamo\Vle\Entity\Course;
 use Kuusamo\Vle\Validation\CourseValidator;
 use Kuusamo\Vle\Validation\ValidationException;
 
+use Doctrine\DBAL\Exception\UniqueConstraintViolationException;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
 
@@ -45,6 +46,8 @@ class AdminDashboardController extends AdminController
                 );
             } catch (ValidationException $e) {
                 $this->alertDanger($e->getMessage());
+            } catch (UniqueConstraintViolationException $e) {
+                $this->alertDanger('URI already in use');
             }
         }