+++ src/Controller/Api/UsersApiController.php
@@ -6,6 +6,7 @@ use Kuusamo\Vle\Entity\User;
 use Kuusamo\Vle\Validation\UserValidator;
 use Kuusamo\Vle\Validation\ValidationException;
 
+use Doctrine\DBAL\Exception\UniqueConstraintViolationException;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
 
@@ -27,14 +28,14 @@ class UsersApiController extends ApiController
         try {
             $validator = new UserValidator;
             $validator($user);
+
+            $this->ci->get('db')->persist($user);
+            $this->ci->get('db')->flush();
         } catch (ValidationException $e) {
             return $this->badRequest($response, $e->getMessage());
+        } catch (UniqueConstraintViolationException $e) {
+            return $this->badRequest($response, 'User already exists');
         }
-
-        // @todo What happens if the user already exists?
-
-        $this->ci->get('db')->persist($user);
-        $this->ci->get('db')->flush();
 
         return $response->withJson($user);
     }